Enable non-integer simulation times

+++ Control_Model_Testing/source/sim/sim.cpp
@@ -4,7 +4,7 @@
 /*
  Runs the simulation.
  */
-int run_sim(int sim_time, double input_time, double dt, int steps,
+int run_sim(double sim_time, double input_time, double dt, int steps,
             double *u, double u_old, double *x_sc, double *x_t,
             double *x_sc_mod, double *xdot_sc_store, double *xdot_t_store,
             double *xdot_sc_mod_store, double **x_sc_store, double **x_t_store,
